Patch yafolding feature:
Since `yafolding` is based on buffer indention detected, thus some
non-ruled occasions whose indentions is not clearly fully for some
case as the docstring of elisp code context will interrupt the context
indention status, thus adding
`entropy/emacs-structure--yafolding-jumping-modes` to using the native
hideshow feature to apply on.

For the perfect idea for code fold and expand, using mode-based
inidividually method treatment was the one picked up, but this may
using the more extensible core library `origami` to extend for, thus
the optional for `entropy/emacs-code-folding-type` is on the way.