`company--perform' patcher bug temporally fixed
With TODO:

EEMACS_BUG: idle trigger not support to
`company-pseudo-tooltip-unless-just-one-frontend' so that we leave it
as original post. This may be corresponding to the the overlay update
mechanism of this frontend's 'UPDTE' command, TODO for while we should
create our own new pseudo tooltip frontend with delay trigger support
replace it.