Use `utf-8-emacs-unix' instead of `utf-8-auto'
Reflect to
https://github.com/jwiegley/emacs-async/issues/165#issue-1562528367

Where simplified quotes as:

I'm told that async.el uses the utf-8-auto coding-system to encode
stuff, assuming that the "auto" part means this coding-system handles
the end-of-line (EOL) format automagically.

This is a mistake. Please read the doc string of utf-8-auto, and you
will see that tha "auto" part is about the BOM, not about the EOL
format. Moreover, on encoding utf-8-auto always produces a BOM,
something that many Lisp (and non-Lisp) programs don't expect at
all. (Due to a bug, utf-8-auto was not producing a BOM on encoding
until now, but Emacs 29 fixes that bug.)

So my suggestion is to replace utf-8-auto with utf-8. The latter
actually does decode the EOL as you'd expect, and is what you usually
want.

If this is used to communicate between two instances of async.el, then
I recommend utf-8-emacs-unix. That is the encoding used by Emacs
internally, and it can represent any character that Emacs is capable
of processing.